Move course search to cours view controller
As space in the tab bar is precious the course search moved to the navigation
bar of the course view controller. The search functionality is important but is not
used very often during the semester. now there is space to put some more often
used functionalities in the tab bar.